requirements.txt
addict==2.4.0
pandas==2.2.3
pillow==11.0.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3